Move server to TS (first step, dirty)
