+++ back/api/package.json
@@ -4,7 +4,7 @@
   "description": "Base de données et API du projet Carton",
   "main": "index.js",
   "scripts": {
-    "start": "node index.js",
+    "start": "tsc && node dist/index.js",
     "test": "mocha",
     "testd": "sudo docker-compose -f docker-compose.yml -f ./test/docker-compose.yml up --build --abort-on-container-exit"
   },
@@ -15,6 +15,9 @@
   },
   "license": "MIT",
   "dependencies": {
+    "@types/express": "^4.17.8",
+    "@types/mongoose": "^5.7.36",
+    "@types/node": "^14.6.4",
     "argon2": "^0.26.2",
     "connect-redis": "^4.0.4",
     "cors": "^2.8.5",
@@ -24,6 +27,7 @@
     "mongoose": "^5.9.14",
     "morgan": "^1.10.0",
     "redis": "^3.0.2",
+    "typescript": "^4.0.2",
     "uuid": "^8.2.0"
   },
   "devDependencies": {
